Accessibility page proofreading
Some comments after proofreading the accessibility page:

The second paragraph ends with a reminder on our duty as a research publisher. I though it is a good chance to also remind that accessibility is essential to effectively fulfill our motto ("opening doors to discovery").

Then, the color section includes a heading with the text "Use High-Contrast Colors". Maybe we don't want to say that, because high contrast colors are not recommended when designing for autistic users. Maybe it's worth replacing that text with "Use Enough Contrasting Colors".

Finally, in the last paragraph, I added a reminder to underline links, since identifying links often relies on color alone.